rsh-chatbot/requirements.txt
langchain>=0.0.267
langchain-community>=0.3.24
pinecone-client>=2.2.1
python-dotenv>=1.0.0
sentence-transformers>=4.1.0
requests>=2.28.0
